+++ R/ProjectionBose.R
@@ -1,134 +1,168 @@
 #' Function to create a strength 2 CA by projection of a Bose OA
 #'
-#' A strength 2 CA in q+1 (q-1)-level columns and q^2-3 runs is created
-#' from a Bose array arranged as SCA
+#' A strength 2 CA with q^2-c runs with q+1 (q-c)-level columns and c s-level columns is created from a q-level Bose array arranged as SCA
 #'
 #' @rdname projectionBose
 #'
+#' @aliases projBoseCA
 #' @aliases projectionBose
 #' @aliases N_projectionBose
 #' @aliases k_projectionBose
 #'
-#' @usage projectionBose(q, fixNA=TRUE, ...)
-#' @usage N_projectionBose(k=NULL, v=NULL, ...)
-#' @usage k_projectionBose(N, v=NULL, ...)
+#' @usage projBoseCA(k, v, fixNA=TRUE, ...)
+#' @usage projectionBose(q, c=1, s=q-c, fixNA=FALSE, ...)
+#' @usage N_projectionBose(k, v, cmax=Inf, ...)
+#' @usage k_projectionBose(N, v, ...)
 #'
-#' @param q a prime power
-#' @param fixNA logical; if FALSE, keeps flexible values; if TRUE, the first \code{q}-3 rows are made constant.
 #' @param k \code{NULL} or requested number of columns (\code{q+1}); if \code{NULL}, inferred from \code{v} as \code{k=v+2}; if specified, \code{k}-1 must be a prime power, or \code{k} will be increased to the nearest such number.
 #' @param v \code{NULL} or number of levels; if \code{NULL}, inferred from \code{k} as \code{v=k+2} or from \code{N} as \code{sqrt(N+3)-1}; if \code{k} is also specified, it takes precedence.
+#' @param q a prime power
+#' @param c an integer number, 0 <= c <= \code{q-2}; for \code{c=0}, the unchanged array created by \code{\link{SCA_Bose}} is returned.
+#' @param s an integer number, 2 <= s <= \code{q-c}; for \code{c=0}, \code{s} is irrelevant.
+#' @param fixNA logical; if FALSE, keeps flexible values; if TRUE, flexible values are arbitrarily fixed.
 #' @param N maximum acceptable number of runs
+#' @param cmax the maximum number permitted for \code{c}; setting \code{cmax} prevents ridiculous values of \code{N} for cases
+#' for which \code{k} and \code{v} are too far apart
 #' @param ... currently not used
 #'
-#' @details
-#' Colbourn (2008) proposed various variants of projection constructions.
-#' This particular variant reduces the number of levels by one and the number of runs by 3,
-#' without increasing the number of columns. It creates CA(q^2-3, 2, q+1, q-1).
-#' The reduction in rows is achieved by using a graph coloring approach.
+#' @returns \code{projBose} returns a covering array in \code{k} columns at \code{v} levels each with \code{v} constant rows,
+#' if this is reasonably possible by projection of a strength 2 Bose OA, which should be the case whenever
+#' \code{v<k} and \code{k-v} not larger than seven and \code{v+(k-v-1)/2} a prime power.\cr
+#' @returns \code{projectionBose} returns a (\code{q^2-c} x \code{q+1+c} matrix with the first \code{q+1} columns
+#' in \code{q-c} levels (labeled from 0 to \code{q-c-1}) and the last \code{c} columns in \code{s} levels (labeled from 0 to \code{s-1})
+#' that is a covering array of strength 2.\cr
+#' \code{N_projectionBose} and \code{k_projectionBose} return a named vector with elements \code{N}, \code{k}, \code{kmax}, \code{v}, \code{q} and \code{c}.
 #'
-#' @returns \code{projectionBose} returns a (\code{q^2-3} x \code{q+1} matrix with columns in \code{q}-1 levels labeled from 0 to q-2 that is a covering array of strength 2.\cr
-#' \code{N_projectionBose} and \code{k_projectionBose} return a named vector with elements \code{N}, \code{k}, \code{v} and \code{q}.
+#' @references Colbourn (2008, Theorem 2.3), Torres-Jimenez et al. (2019, Fig. 25)
 #'
 #' @examples
-#' # create a CA(22,2,6,4)
-#' projectionBose(5)  ## two constant rows
-#' projectionBose(5, fixNA=FALSE) ## two flexible values
+#' # create a CA(24,2,7,4)
+#' projectionBose(5)  ## four constant rows, some flexible values
+#' projectionBose(5, fixNA=TRUE) ## flexible values arbitrarily filled
+#' projectionBose(5, c=2, s=2) ## 6 3-level columns (3=5-2), two 2-level columns
 #'
-#' # create a CA(46,2,8,6)
+#' # create a CA(48,2,9,6)
 #' dim(D <- projectionBose(7))
 #' coverage(D, 2)
-#' eCAN(2,8,6) ## four runs difference
+#' eCAN(2,9,6) ## two runs difference
 #'
-#' # create a CA(61,2,9,7)
-#' dim(projectionBose(8))
-#' eCAN(2, 9, 7)  ## two runs difference
+#' # create a CA(63,2,9,7)
+#' dim(projBoseCA(9,7))
+#' eCAN(2, 9, 7)  ## four runs difference
 #'
-#' # create a CA(166,2,14,12)
-#' dim(projectionBose(13))
-#' eCAN(2, 14, 12)  ## one run difference
+#' # create a CA(166,2,15,12)
+#' dim(projBoseCA(15,12))
+#' eCAN(2, 15, 12)  ## optimal
 #'
 #' # querying k or N
-#' N_projectionBose(k=21)
-#' eCAN(2, 24, 22) ## one run difference
-#' eCAK(2, 526, 22) ## as good as it gets
-#'
-#' N_projectionBose(v=6)
-#' k_projectionBose(N=50)
-#'
-#' k_projectionBose(N=118)
-#' eCAN(2,12,10) ## two runs difference
+#' N_projectionBose(k=21, v=15) ## not a good use case
+#' eCAN(2, 21, 15) ## the above design is the basis
+#' eCAK(2, 357, 15) ## four more columns
 #'
 
 #' @export
-projectionBose <- function(q, fixNA=TRUE, ...){
-  stopifnot(is.logical(fixNA))
-  ## first row immediately dropped
-  arr <- SCA_Bose(q)[-1,]
-  arr[arr==0] <- NA ## removing the level 0 entirely
-  ## arrsub has the first rows constant in increasing order
-  arrsub <- arr[,-(q+1)]
-
-  vertices <- which(is.na(arrsub), arr.ind=TRUE)
-  ## store the NA-positions, which are graph vertices
-
-  G <- igraph::make_empty_graph(n=nrow(vertices), directed = FALSE)
-
-  ## find edges
-  paare <- nchoosek(q, 2)
-  for (paar in 1:ncol(paare)){
-    spalten <- paare[,paar]
-    for (i in 1:2){
-      ## loop over the to-be-deleted rows
-      ## assign edges as indicated in Colbourn 2008 (proof of Lemma 3.2)
-      row1 <- which(arrsub[,spalten[1]]==i & is.na(arrsub[,spalten[2]]))
-      row2 <- which(is.na(arrsub[,spalten[1]]) & arrsub[,spalten[2]]==i)
-      from <- which(vertices[,1]==row1 & vertices[,2]==spalten[2])
-      to <- which(vertices[,1]==row2 & vertices[,2]==spalten[1])
-      G <- igraph::add.edges(G, c(from, to))
-    }
-  }
-  ## yields numbers 1 and 2, perfect for being assigned to
-  ## NA values
-  farben <- igraph::greedy_vertex_coloring(G)
-  ## suitably assign values in order to achieve coverage
-  for (i in 1:nrow(vertices)){
-    arrsub[vertices[i,1], vertices[i,2]] <- farben[i]
-  }
-  aus <- cbind(arrsub[-(1:2),], arr[-(1:2),q+1])-1
-  if (fixNA) aus[1:(q-3), q+1] <- aus[1:(q-3), 1]
+projBoseCA <- function(k, v, fixNA=TRUE, ...){
+  Call <- sys.call()
+  hilf <- N_projectionBose(k, v)  ## with default cmax=Inf
+  if (is.na(hilf[1])) stop("This construction is not available for this combination of k and v.")
+  aus <- projectionBose(hilf["q"], hilf["c"], fixNA=fixNA, ...)[,1:k]
+  class(aus) <- c("ca", class(aus))
+  attr(aus, "PCAstatus") <- list
+  attr(aus, "call") <- Call
+  attr(aus, "origin") <- "projection (Colbourn 2008 Thm 2.3)"
   aus
 }
 
 #' @export
-N_projectionBose <- function(k=NULL, v=NULL, ...){
-  primpotenzen <- primedat$q
-  if (is.null(k) && is.null(v)) stop("k and v must not both be NULL")
-  if (!is.null(k)){
-    stopifnot(is.numeric(k))
-    stopifnot(k %% 1==0)
-    if (!(k-1) %in% primedat$q) {
-      k <- min(primpotenzen[which(primpotenzen>=k)])+1
-      message("k was increased to nearest usable value")
+projectionBose <- function(q, c=1, s=q-c, fixNA=FALSE, ...){
+  ## s is the number of levels of the additional columns
+  ## if it equals q-c, the resulting CA is uniform
+  stopifnot(is.numeric(q), is.numeric(c), is.numeric(s))
+  stopifnot(c>=0)
+  stopifnot(c<=q-2)    ## make sure at least two levels remain
+  stopifnot(s <= q-c)  ## make sure that s is permissible
+  stopifnot((c%%1)==0)
+  stopifnot(is.logical(fixNA))
+  arr <- SCA_Bose(q)                  ## also checks for appropriate q
+  if (c==0) return(arr)
+  newmat <- cbind(arr, matrix(NA, nrow(arr), c))
+       ## c added columns, initialized as flex
+
+  ## immediately use final symbols only for the first s columns
+  replacesel <- setdiff(0:(q-1), 0:(c-1))
+
+  for (j in 1:c){
+    ## loop through the first c rows (rows to be removed)
+
+    for (k in 1:s){
+      ## determine C_k
+      hilf <- setdiff(which(newmat[,k]==newmat[j,k]), (1:c))
+      ## fill the new column
+      newmat[hilf, q+1+j] <- k - 1 + c  ## immediately correct levels
+                                        ## no after-treatment needed
+                                        ## except subtracting c
+      ## treat the first s columns
+      if (length(replacesel) < length(hilf))
+        replace <- sample(
+          c( replacesel, rep(NA, length(hilf)-length(replacesel)) )
+          )
+      else replace <- replacesel
+             ## permutation of target values (starting with c)
+      newmat[hilf, k] <- replace
     }
-    if (!is.null(v)) if (!k==v+2){
-      message("v was adapted to be compatible with k")
-    }
-    v <- k - 2
-    ## check for prime
-    q <- k-1
-  }else{
-    stopifnot(is.numeric(v))
-    stopifnot(v %% 1==0)
-    ## check for prime
-    if (!(v+1) %in% primedat$q) stop("v+1 must be a prime power")
-    q <- v+1
-    k <- q+1
+
+    ## treat value j-1 in columns s+1 to q+1 (wild card)
+    newmat[(c+1):(q^2),(s+1):(q+1)][which(newmat[(c+1):(q^2),(s+1):(q+1)]==j-1)] <- NA
   }
-  c(N=q^2-3, k=k, v=v, q=q)
+
+
+  aus <- newmat[-(1:c),] - c
+  ## make first rows constant
+  for (i in 1:q){
+    if (s==q-c || i <= s)
+      aus[i, (q+1):(q+1+c)] <- i-1
+    else
+      aus[i, (q+1)] <- i-1
+  }
+  missings1 <- which(is.na(aus[,1:(q+1)]))
+  missings2 <- which(is.na(aus[,(q+2):(q+1+c)]))
+  if (fixNA){
+    aus[,1:(q+1)][missings1] <- sample(0:(q-c-1), length(missings1), replace = TRUE)
+    aus[,(q+2):(q+1+c)][missings2] <- sample(0:(s-1), length(missings2), replace = TRUE)
+  }
+  aus
+}
+
+#' @export
+N_projectionBose <- function(k, v, cmax=Inf, ...){
+  ## k and v must be given
+  ## k is a minimum, v must be matched exactly
+  ## there must be a prime power q with v+c=q and k<=q+1+c
+  primpotenzen <- primedat$q
+    stopifnot(is.numeric(k), is.numeric(v), is.numeric(cmax))
+    stopifnot(k %% 1==0, v%%1==0)
+    stopifnot(v < k)
+    kalt <- k
+    if (!((k-v)%%2)==1){
+      k <- k+1
+      }
+    c <- (k-v-1)/2
+    q <- NA
+    while(v+c < k){
+    if ((v+c) %in% primpotenzen){
+      q <- v+c
+      break
+      }else
+      c <- c+1
+      }
+    if (is.na(q)) return(NA) ## no adequate scenario found
+    if (c > cmax) return(NA) ## no adequate and permissible scenario found
+  c(N=q^2-c, k=kalt, kmax=q+1+c, v=v, q=q, c=c)
 }
 
 #' @export
 k_projectionBose <- function(N, v=NULL, ...){
+  ## not yet implemented
   primpotenzen <- primedat$q
   hilf <- primpotenzen^2 - 3
   if (!(N %in% (primpotenzen^2-3))){